Move shares listing to a service
